+++ example/lib/main.dart
@@ -16,6 +16,7 @@ class _MyAppState extends State<MyApp> {
   @override
   Widget build(BuildContext context) {
     return ToolboxApp(
+      noItemsFoundWidget: Icon(Icons.subject),
       child: MaterialApp(
         localizationsDelegates: const [
           toolbox.S.delegate,