+++ lib/src/config/toolbox_app.dart
@@ -1,18 +1,26 @@
 import 'dart:collection';
 
 import 'package:flutter/widgets.dart';
+import 'package:flutter_toolbox/src/config/toolbox_config.dart';
+import 'package:flutter_toolbox/src/ui/listview/pagewise/flutter_pagewise.dart';
 import 'package:oktoast/oktoast.dart';
+import 'package:provider/provider.dart';
 
 LinkedHashMap<_ToolboxAppState, BuildContext> contextMap = LinkedHashMap();
 
 class ToolboxApp extends StatefulWidget {
-  /// Usually should be [MaterialApp] or [CupertinoApp].
-  final Widget child;
-
   const ToolboxApp({
     Key key,
     @required this.child,
+    this.noItemsFoundBuilder,
+    this.noItemsFoundWidget,
   }) : super(key: key);
+
+  /// Usually should be [MaterialApp] or [CupertinoApp].
+  final Widget child;
+
+  final NoItemsFoundBuilder noItemsFoundBuilder;
+  final Widget noItemsFoundWidget;
 
   @override
   _ToolboxAppState createState() => _ToolboxAppState();
@@ -38,7 +46,13 @@ class _ToolboxAppState extends State<ToolboxApp> {
       textPadding: EdgeInsets.symmetric(horizontal: 32, vertical: 16),
       position: ToastPosition.bottom,
       radius: 50,
-      child: widget.child,
+      child: Provider.value(
+        value: ToolboxConfig(
+          noItemsFoundBuilder: widget.noItemsFoundBuilder,
+          noItemsFoundWidget: widget.noItemsFoundWidget,
+        ),
+        child: widget.child,
+      ),
     );
   }
 }